chore: moved components into specific directory
