+++ app/src/features/surveys/observations/sampling-sites/site/accordion-details/period/SamplingSiteListPeriod.tsx
@@ -7,7 +7,7 @@ import Typography from '@mui/material/Typography';
 import { IObservationsContext } from 'contexts/observationsContext';
 import { IObservationsPageContext } from 'contexts/observationsPageContext';
 import dayjs from 'dayjs';
-import { ImportObservationsButton } from 'features/surveys/observations/ImportObservationsButton';
+import { ImportObservationsButton } from 'features/surveys/observations/components/ImportObservationsButton';
 import { GetSamplingPeriod } from 'interfaces/useSamplingPeriodApi.interface';
 
 interface ISamplingSiteListPeriodProps {
